Thoroughly updated the main function
The main function now has pretty much everything you need to recreate the figures we are going to
use for the paper. It reads in the original information, splits things up into groups, performs permutations,
plots rates of recombination, plots the rates of recombination between real and sim, calculates EHH and REHH
in the permuted simulations at distances between 50K and 500K, and finally plots the real versus simulated
data in a cloud/heat map.